Change project name and dependencies in composer.json
Due to the transfer of the project from 'convenia/excel' to 'lunia/excel', changes were made in the composer.json file. The 'excel-manager' dependency was updated from a specific version ("^1.9") to any version ("dev-master") to indicate compatibility with any version of 'excel-manager'. Also, a link to the new version control repository was added under "repositories".